Extract function to external.commands
