+++ perun/utils/external/meson.build
@@ -2,6 +2,7 @@ perun_external_dir = perun_utils_dir / 'external'
 
 perun_external_files = files(
     '__init__.py',
+    'commands.py',
     'environment.py',
 )
